+++ packages/stash/lib/src/api/cache/cache_store.dart
@@ -1,0 +1,6 @@
+import '../store.dart';
+import 'cache_entry.dart';
+import 'cache_info.dart';
+
+// Cache store marker interface
+abstract class CacheStore extends Store<CacheInfo, CacheEntry> {}